chore: move hsm generatated header
