+++ .github/workflows/quom.yml
@@ -24,5 +24,5 @@ jobs:
 
     - name: Check Single Header File
       run: |
-        quom --include_directory src/include/ src/include/hsm/hsm.h src/include/hsm/hsm_gen.h
+        quom --include_directory src/include/ src/include/hsm/hsm.h src/include/hsm/gen/hsm.h
         git diff --exit-code
