+++ src/include/hsm/gen/.clang-format
@@ -1,0 +1,1 @@
+DisableFormat: true
